refactor: added use-user-state hook for profile
