+++ common/client/fragments/profile-fragment/src/profile.component.tsx
@@ -1,9 +1,6 @@
 import React                  from 'react'
 import { FC }                 from 'react'
 import { FormattedMessage }   from 'react-intl'
-import { useRouter }          from 'next/router'
-import { useEffect }          from 'react'
-import { useState }           from 'react'
 
 import { User }               from '@shared/data'
 import { Button }             from '@ui/button'
@@ -15,36 +12,15 @@ import { Layout }             from '@ui/layout'
 import { Text }               from '@ui/text'
 import { Title }              from '@ui/title'
 import { WrapperWhite }       from '@ui/wrapper'
-import { mockAuthUserId }     from '@shared/data'
-import { useGetUser }         from '@shared/data'
 
 import { ButtonEditProfile }  from './button-edit-profile'
 import { ProfileAvatar }      from './profile-avatar'
 import { ProfileDescription } from './profile-description'
 import { ProfileProjects }    from './profile-projects'
+import { useUserState }       from './hooks'
 
 export const Profile: FC = () => {
-  const router = useRouter()
-  const queryId = typeof router.query.id === 'string' ? router.query.id : ''
-
-  const [user, setUser] = useState<User>()
-  const [isMyProfile, setIsMyProfile] = useState(false)
-
-  const { user: fetchedUser, teams, projects } = useGetUser({ id: queryId, skip: 0, take: 20 })
-
-  useEffect(() => {
-    if (fetchedUser) {
-      const { avatarUrl, description, id, tags } = fetchedUser
-
-      if (!avatarUrl || !description || !tags || tags.length === 0) {
-        router.push('/onboarding')
-      }
-
-      if (id === mockAuthUserId) setIsMyProfile(true)
-
-      setUser(fetchedUser)
-    }
-  }, [router, fetchedUser])
+  const { user, setUser, isMyProfile, projects, teams } = useUserState()
 
   const handleEditUser = (editedUser: User) => setUser(editedUser)
